Add first draft of meta-ros maintainer notes
These notes are meant to describe the workflow that a maintainer
would use to add a new Yocto Release or ROS Distro.  They also
include instructions on how to do an automated build for testing.

The notes are based on the work done previously by Martin Jansa.
The hope it to preserve the well-documented commit history to
make it easy to manage going forward.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>